Many changes
Now using Amazon web services database. Can update and reach the database. Changed structure of get_events output in order to easily add the information to the database. Adjusted kivy_app to be able to display the information despite the new structure. It also will not crash if the user has no internet connection, but display a message to indicate to the user of no connection.