Button symbol changes
A first implementations of the new art (by Robert Morgan). The music
on/off button now has an icon for on, off and in between. The back
button now also has an icon. The music button however does not animate
between it's images fluidly. The search and refresh icons have not been
used yet in this commit.