File Organisation and Button Symbols
First, many of the files were moved in order to improve the structure in
a more logical way. Second, more buttons recieved symbols and their text
was removed. Lastly, an animation very similar to the refresh animation
now exists for the searching action. When the text is no longer visible
(the short gap of time) the search icon fades into sight before fading
out to opacity = 0.