Big documentation commit!
Every method now has a docstring describing what it does, the
parameters it expects and it's returned value. There is only one
exception as it is a @property and adding a docstring caused failures
throughout the app (where the property was used). In addition, single
line comments were added under if __name__ == '__main__' to better
describe each main action that is taken in the start up and termination
of the app.