Final commit, updated README and minor changes
README.md was updated with proper instructions if someone else would
like to run this on their own machine. Minor changes were made to the
small arrow buttons. The kivy_app.py file had some minor changes to do
with sys.path. The app was developed within PyCharmCE, which enabled me
to set the src directory as a parent folder. To ensure everything works
outside of PyCharm, the src path is manually added to sys.path. The
default of the music_active boolean in user_settings.json should be
true, but while developing, I set it to false. Now it is true by default.